Added a view messages page and added some minor changes on sign-up.php
Added a message page where a signed in user can view all sent messages.

Added some functions to dbh.inc.php

made a small change so that the input of email on the sign-up page must have the syntax of something"@"somethingelse

Also added a link to the msg page and the home page.

Changed some table column names so that their all in english and better :)